refactor: delete gemini.test.ts, move signal/TTY tests to cleanup.test.ts
Addresses review feedback: removed the new gemini.test.ts file and moved
signal handler and TTY check tests into cleanup.test.ts where they belong.
Reverted the eslint allowDefaultProject workaround since deleting the .ts
file eliminates the stem conflict.
